Modity post list view controller to display the date of a post in simple form and segue to post detail view controller on tap.
Add post detail view controller to review the post and allow the user to segue to the tones scene.

Analyze the tone of a post when it is entered into Core Data and create a tone entity, associating it with the post entity.

Hide the cancel and post to facebook buttons, and display the back buttone, on the tones view controller when the segue originated from the post detial view controller. Do the opposite when the segue originated from the post to facebook scene.